JS date object & formatted values supported
* added the ability to specify a JS date object as data value
* added the ability to specify an array, so you can customize the data value 'v' and the formatted version 'f'

https://developers.google.com/chart/interactive/docs/reference

v [Optional] The cell value. The data type should match the column data type. If null, the whole object should be empty and have neither v nor f properties.

f [Optional] A string version of the v value, formatted for display. The values should match, so if you specify Date(2008, 0, 1) for v, you should specify "January 1, 2008" or some such string for this property. This value is not checked against the v value. The visualization will not use this value for calculation, only as a label for display. If omitted, a string version of v will be used.